Expose some functions outside of SessionManager
