+++ src/presentation/authentication/mdoc.rs
@@ -2,11 +2,12 @@ use crate::cbor;
 use crate::cose;
 use crate::definitions::device_response::Document;
 use crate::definitions::issuer_signed;
+use crate::definitions::session::SessionTranscript;
 use crate::definitions::x509::X5Chain;
 use crate::definitions::DeviceAuth;
 use crate::definitions::Mso;
 use crate::definitions::{
-    device_signed::DeviceAuthentication, helpers::Tag24, SessionTranscript180135,
+    device_signed::DeviceAuthentication, helpers::Tag24,
 };
 use crate::presentation::reader::Error;
 use anyhow::Result;
@@ -30,10 +31,12 @@ pub fn issuer_authentication(x5chain: X5Chain, issuer_signed: &IssuerSigned) -> 
         .map_err(Error::IssuerAuthentication)
 }
 
-pub fn device_authentication(
+pub fn device_authentication<S>(
     document: &Document,
-    session_transcript: SessionTranscript180135,
-) -> Result<(), Error> {
+    session_transcript: S,
+) -> Result<(), Error>
+where
+S: SessionTranscript + Clone{
     let mso_bytes = document
         .issuer_signed
         .issuer_auth
